refactor(util): replaced default arm with panic
as mentioned in #1, returning the root dir path if no operating system matches is a possible cause for errors. To fix this opus now panics if the os isnt matched with linux, macos or windows. To circumvent this as a user, set the 'OPUS_PATH' variable to point opus to the correct directory and database. If the parent dir of the database isnt found, opus creates it as well as the database file itself (This can and will be probably changed in the future).